Finalizing moving all commands to new command system
