+++ src/main/java/usagibot/twitch/TwitchClient.java
@@ -1,27 +1,19 @@
 package usagibot.twitch;
 
 import com.github.philippheuer.credentialmanager.domain.OAuth2Credential;
-import com.github.philippheuer.events4j.core.EventManager;
 import com.github.philippheuer.events4j.simple.SimpleEventHandler;
 import com.github.twitch4j.TwitchClientBuilder;
 import lombok.extern.slf4j.Slf4j;
 import usagibot.UsagiBot;
-import usagibot.twitch.commands.Command;
-import usagibot.twitch.commands.CommandClient;
-import usagibot.twitch.commands.CommandEvent;
-import usagibot.twitch.commands.twitchcommands.HelpCommand;
-import usagibot.twitch.commands.twitchcommands.PongCommand;
-import usagibot.twitch.commands.twitchcommands.RequestToggleCommand;
-import usagibot.twitch.event.ChatEvent;
-
-import java.beans.EventHandler;
+import usagibot.commands.CommandClient;
+import usagibot.commands.twitchcommands.*;
 
 @Slf4j
 public class TwitchClient {
 
     public static com.github.twitch4j.TwitchClient client;
     public static OAuth2Credential credentials = new OAuth2Credential("twitch", UsagiBot.getConfig().getTwitchPassword());
-    CommandClient clientC;
+    CommandClient commandClient;
 
     /**
      * The Twitch IRC client constructor
@@ -52,13 +44,18 @@ public class TwitchClient {
      */
     public void loadListeners() {
         SimpleEventHandler eventHandler = client.getEventManager().getEventHandler(SimpleEventHandler.class);
-        clientC = new CommandClient(eventHandler);
+        commandClient = new CommandClient(eventHandler);
         addCommands();
     }
 
+    /**
+     * Add commands to the list of current avaliable commands
+     */
     public void addCommands() {
-        clientC.addCommand(new PongCommand());
-        clientC.addCommand(new HelpCommand());
-        clientC.addCommand(new RequestToggleCommand());
+        commandClient.addCommand(new HelpCommand());
+        commandClient.addCommand(new NowPlayingCommand());
+        commandClient.addCommand(new RequestToggleCommand());
+        commandClient.addCommand(new StarLimitCommand());
+        commandClient.addCommand(new StatsCommand());
     }
 }
