--- src/main/java/usagibot/twitch/commands/twitchcommands/StarLimitCommand.java
@@ -1,4 +1,0 @@
-package usagibot.twitch.commands.twitchcommands;
-
-public class StarLimitCommand {
-}
